+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# [0.2.1](https://github.com/tjtelan/git-meta-rs/compare/v0.1.0...v0.2.0) (2021-02-12)
+- Check for changes in a path ([#4](https://github.com/tjtelan/git-meta-rs/issues/4))
+- Expand partial commit ids ([#7](https://github.com/tjtelan/git-meta-rs/issues/7))
+- Added some tests and examples using this repo's commits
 # [0.2.0](https://github.com/tjtelan/git-meta-rs/compare/v0.1.0...v0.2.0) (2021-02-12)
 - Loosened requirements for local branches resolving to remote branches ([#5](https://github.com/tjtelan/git-meta-rs/issues/5))
 - Modified `GitRepo::with_commit()` to take `Option<Commit>`